add example to webwork sample-chapter.ptx
I spent a lot of time trying to figure out how to do multianswer and also how to replace perl's "and" symbol, so I thought I'd try to save other people the time.  Oscar had a thread on support in August on multianswer, (I borrowed and modified his example), so I think it's of general interest.